@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -45,6 +45,7 @@ dependencies: #dependencias que seran usadas en tiempo de ejecucion
   flutter_rating_bar: ^4.0.1 #Rating progressBar ✦✦✦✧✧
   http: ^0.13.4 #consumo de api rest
   shared_preferences: ^2.0.15 #persistencia
+  elegant_notification: ^1.6.4 #Despliegua una notificacion elegante para informar al usuario
 
 dev_dependencies: #dependencias que seran usadas unicamente en tiempo de compilacion
 
