@@ -7,8 +7,7 @@ import 'package:http/http.dart' as http;
 class RateRequest {
   Future<Map<String, dynamic>> rate(String host, Map<String, String> header, Map<String, String> body) async {
     var url = Uri.parse(host+"/calification/create");
-    var response =
-        await http.post(url, body: jsonEncode(body), headers: header);
+    var response = await http.post(url, body: jsonEncode(body), headers: header);
     // print('Response status: ${response.statusCode}');
     // print('Response body: ${response.body}');
     if (response.statusCode == 200) {
