@@ -1,6 +1,8 @@
-import 'dart:convert';
+// ignore_for_file: file_names
 
-import 'package:agenda_movil/src/Model/SubscriptionModel.dart';
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:agenda_movil/src/Persistence/Percistence.dart';
 
 import 'package:http/http.dart' as http;
@@ -9,23 +11,47 @@ class SubscriptionRequest {
 
   final Percistence _percistence = Percistence();
 
-  Future<Map<String, dynamic>> viewSubscriptions(String host, Map<String, String> header, Map<String, String> body) async {
-    var url = Uri.parse( host+"/subscription/show");
-    
-    var response = await http.post(url, body: jsonEncode(body), headers: header);
-    // print('Response status: ${response.statusCode}');
-    // print('Response body: ${response.body}');
-    if (response.statusCode == 200) {
-      _percistence.subscription = response.body;
+  Future<bool> viewSubscriptions(String host, Map<String, String> header, Map<String, dynamic> body) async {
+    try {
+      var url = Uri.parse( host+"/subscription/show");
+      var response = await http.post(url, body: jsonEncode(body), headers: header);
+      // print('Response body: ${response.body}');
+      if (response.statusCode == 200) {
+        _percistence.subscription = response.body;
+        return true; 
+      }
+    } on SocketException catch (e) {
+      return false;
+    }
+    return null!;
+  }
+
+  Future<Map<String, dynamic>> subscribe(String host, Map<String, String> header, Map<String, dynamic> body) async {
+    try {
+      var url = Uri.parse(host+"/subscription/subscribe");
+      var response = await http.post(url, body: jsonEncode(body), headers: header);
+      // print('Response status: ${response.statusCode}');
+      // print('Response body: ${response.body}');
+      if (response.statusCode == 200) {
+        return {
+          "status": true, // "OK"
+          "message": "El registro ha sido exitoso"
+        };
+      }else if(response.statusCode == 400){
+        return {
+          "status": false, // "ERROR"
+          "type":"info",
+          "message": "Ya te encuentras registrado en esta materia"
+        };
+      }
+    }on SocketException catch(e){
       return {
-        "status": true, // "OK"
+        "status": false, // "ERROR"
+        "type":"error",
+        "message": "No pudimos completar la accion, revisa si cuentas con conexion a internet"
       };
     }
-    return {
-      "status": false, // "ERROR"
-      "emoji": "(⊙_⊙)？",
-      "message": "Revisa tu conexion a internet e intenta de nuevo, no pudimos conextar con los servidores"
-    };
+    return null!;
   }
 
 }
