@@ -3,16 +3,29 @@
 import 'dart:async';
 
 import 'package:agenda_movil/src/API/RateRequest.dart';
-import 'package:agenda_movil/src/API/RegisterRequest.dart';
-import 'package:agenda_movil/src/API/LogingRequest.dart';
+import 'package:agenda_movil/src/API/MatterRequest.dart';
+import 'package:agenda_movil/src/API/StudentRequest.dart';
 import 'package:agenda_movil/src/API/SubscriptionRequest.dart';
 import 'package:agenda_movil/src/Logic/Streams.dart';
+import 'package:agenda_movil/src/Model/SubscriptionModel.dart';
+
+import '../Model/StudentModel.dart';
+import '../Persistence/Percistence.dart';
 
 
 class Management{
-  // String host = "https://back-end-agenda-digital.herokuapp.com";
-  // String host = "http://127.0.0.1:8080";
-  String host = "https://40e2-2800-e2-ab00-26d-2d19-a751-4856-6bd8.ngrok.io";
+
+  Management(){
+    if(_percistence.student!=""){
+      setStudent();
+    }
+    if(_percistence.subscription!=""){
+      setSubscriptionList();
+    }
+  }
+
+  final Percistence _percistence = Percistence();
+  String host = "https://back-end-agenda-digital.herokuapp.com";
   Map<String, String> header = {
     "Accept": "application/json",
     "content-type": "application/json"
@@ -22,23 +35,42 @@ class Management{
 
 //BottonMenuBar
   int _currentIndex = 0;
-
   int get getIndex {
     return _currentIndex;
   }
-
   set setIndex(int index) {
     _currentIndex = index;
   }
 
-  //login
-  Future<Map<String, dynamic>> logingRequest(Map<String, String> body) {
-    return LogingRequest().loggin(host, header, body);
+  late StudentModel? _student;
+  StudentModel get getStudent{
+    return _student!;
+  }
+  void setStudent(){
+    _student = studentFromJson(_percistence.student);
   }
 
-  //register
+  List<SubscriptionModel> _subsciptionList = List.empty(growable: true);
+  List<SubscriptionModel> get getSubscriptionList{
+    return _subsciptionList;
+  }
+  void setSubscriptionList(){
+    _subsciptionList = subscriptionListFromJson(_percistence.subscription);
+  }
+
+
+
+  //Student-Login-Register
+  Future<Map<String, dynamic>> logingRequest() {
+    Map<String, String> body = {
+      "email": streams.email,
+      "password": streams.password
+    };
+    return StudentRequest().loggin(host, header, body);
+  }
+
   Future<Map<String, dynamic>> registerRequest(Map<String, String> body) {
-    return RegisterRequest().register(host, header, body);
+    return StudentRequest().register(host, header, body);
   }
 
   //rate
@@ -47,7 +79,33 @@ class Management{
   }
 
   //subscripciption
-  Future<Map<String, dynamic>> subscripciptionRequest(Map<String, String> body) {
-    return SubscriptionRequest().viewSubscriptions(host, header, body);
+  void subscripciptionRequest() async{
+    Map<String, dynamic> body = {
+      "student":getStudent.toJson()
+    };
+    bool ok = await SubscriptionRequest().viewSubscriptions(host, header, body);
+    if(ok){
+      setSubscriptionList();  
+    }
+  }
+
+  //matter
+  Future<Map<String, dynamic>> createMatterRequest(String rgb){
+    Map<String, dynamic> body = {
+      "name": streams.matterName,
+      "rgb": rgb,
+      "student":getStudent.toJson()
+    };
+    return MatterRequest().create(host, header, body);
+  }
+  
+  Future<Map<String, dynamic>> searchMatterRequest(){
+    Map<String, dynamic> body = {
+      "matter":{
+        "id":streams.matterId
+      },
+      "student":getStudent.toJson()
+    };
+    return SubscriptionRequest().subscribe(host, header, body);
   }
 }
