@@ -4,34 +4,25 @@ import 'dart:convert';
 
 import 'MatterModel.dart';
 
-List<SubscriptionModel> subscriptionFromJson(String str) => List<SubscriptionModel>.from(json.decode(str).map((x) => SubscriptionModel.fromJson(x)));
-
-String subscriptionToJson(List<SubscriptionModel> data) => json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
+List<SubscriptionModel> subscriptionListFromJson(String str) => List<SubscriptionModel>.from(json.decode(str).map((x) => SubscriptionModel.fromJson(x)));
 
 class SubscriptionModel {
-    SubscriptionModel({
-        required this.id,
-        required this.matter,
-        required this.student,
-        required this.request,
-    });
+  SubscriptionModel(
+    {
+      required this.id,
+      required this.matter,
+      required this.request,
+    }
+  );
 
-    final int id;
-    final MatterModel matter;
-    final dynamic student;
-    final bool request;
+  final int id;
+  final MatterModel matter;
+  final bool request;
 
-    factory SubscriptionModel.fromJson(Map<String, dynamic> json) => SubscriptionModel(
-        id: json["id"],
-        matter: MatterModel.fromJson(json["matter"]),
-        student: json["student"],
-        request: json["request"],
-    );
+  factory SubscriptionModel.fromJson(Map<String, dynamic> json) => SubscriptionModel(
+    id: json["id"],
+    matter: MatterModel.fromJson(json["matter"]),
+    request: json["request"],
+  );
 
-    Map<String, dynamic> toJson() => {
-        "id": id,
-        "matter": matter.toJson(),
-        "student": student,
-        "request": request,
-    };
 }
